Add common Information css module
- removed redundant Information css function
- changed to use common Information css module
