+++ src/presentational/EntranceCheckForm.jsx
@@ -6,14 +6,10 @@ import InputPart from './InputPart';
 
 import Button from '../style/Button';
 
+import Information from '../style/Information';
+
 const Wrapper = styled.div(() => ({
   margin: '10px 0',
-}));
-
-const Information = styled.p(() => ({
-  margin: '0 15px',
-  fontSize: '12px',
-  color: '#A4A4A4',
 }));
 
 export default function EntranceCheckForm({
