+++ src/style/Information.jsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const Information = styled.p(() => ({
+  margin: '5px 15px',
+  fontSize: '12px',
+  color: 'gray',
+  textAlign: 'center',
+}));
+
+export default Information;
